Move interface-related files into a separate folder.
