+++ console/src/app/core/models/chart/mongoose-chart-interface/mongoose-chart-data-provider.interface.ts
