+++ console/src/app/core/models/chart/mongoose-chart-interface/mongoose-chart-options.ts
